Created the endpoint to read from csv, need to organize the code following clean architecture

+++ app/models/character.go
@@ -1,6 +1,6 @@
 package models
 
 type Character struct {
-	Id   uint64 `json:"id"`
+	Id   int    `json:"id"`
 	Name string `json:"name"`
 }
